% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clip_matrix_values}
\alias{clip_matrix_values}
\title{Clip matrix values to specified range}
\usage{
clip_matrix_values(mat, clip_range)
}
\arguments{
\item{mat}{Matrix to clip}

\item{clip_range}{Numeric vector of length 2 with min and max values}
}
\value{
Matrix with values clipped to range
}
\description{
Clip matrix values to specified range
}
