% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{Denoise data by setting all latent factor to their median values and reversing the regression model}
\usage{
denoise(x, data = "y", cell_attr = x$cell_attr, do_round = TRUE,
  do_pos = TRUE)
}
\arguments{
\item{x}{A list that provides model parameters and optionally meta data; use output of vst function}

\item{data}{The name of the entry in x that holds the data}

\item{cell_attr}{Provide cell meta data holding latent data info}

\item{do_round}{Round the result to integers}

\item{do_pos}{Set negative values in the result to zero}
}
\value{
De-noised data as UMI counts
}
\description{
Denoise data by setting all latent factor to their median values and reversing the regression model
}
