% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate data from regularized models.}
\usage{
generate(vst_out, data = "y", genes = rownames(vst_out$model_pars_fit),
  cell_attr = x$cell_attr, background_only = FALSE, do_round = TRUE,
  do_pos = TRUE, make_sparse = FALSE)
}
\arguments{
\item{vst_out}{A list that provides model parameters and optionally meta data; use output of vst function}

\item{data}{The name of the entry in x that holds the data (Pearson residuals)}

\item{cell_attr}{Provide cell meta data holding latent data info}

\item{background_only}{Set to TRUE to only generate the background signal (do not add back residuals)}

\item{do_round}{Round the result to integers}

\item{do_pos}{Set negative values in the result to zero}

\item{make_sparse}{Make the output a dgCMatrix (if do_pos and do_round are also TRUE)}
}
\value{
Generated data
}
\description{
Generate data from regularized models. If background_only is set to TRUE no residuals are
added to the simulated data.
}
\examples{
vst_out <- vst(pbmc, return_cell_attr = TRUE, res_clip_range = c(-Inf, Inf))
generated_data <- generate(vst_out)

}
