% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nz_median}
\alias{get_nz_median}
\title{Get median of non zero UMIs from a count matrix using a subset of genes (slow)}
\usage{
get_nz_median(umi, genes = NULL)
}
\arguments{
\item{umi}{Count matrix}

\item{genes}{List of genes to calculate statistics. Default is NULL which returns the non-zero median using all genes}
}
\value{
A numeric value representing the median of non-zero entries from the UMI matrix
}
\description{
Get median of non zero UMIs from a count matrix using a subset of genes (slow)
}
