% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x, eps = 1)
}
\arguments{
\item{x}{array}

\item{eps}{small value to add to x to avoid log(0); default is 1}
}
\value{
geometric mean
}
\description{
Geometric mean
}
