% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmean_sparse}
\alias{gmean_sparse}
\title{Geometric mean for sparse data}
\usage{
gmean_sparse(x, N, eps = 1)
}
\arguments{
\item{x}{array of non-zero values}

\item{N}{total number of values (including zeros)}

\item{eps}{small value to add to x (and zeros) to avoid log(0); default is 1}
}
\value{
geometric mean
}
\description{
Geometric mean for sparse data
}
