% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify outliers}
\usage{
is_outlier(y, x, n, th = 40)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Independent variable}

\item{n}{Number of observations per bin}

\item{th}{Outlier score threshold}
}
\value{
Boolean vector
}
\description{
Identify outliers
}
