% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.sparse}
\alias{make.sparse}
\title{Convert a given matrix to dgCMatrix}
\usage{
make.sparse(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
A dgCMatrix
}
\description{
Convert a given matrix to dgCMatrix
}
