% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_regressor_data}
\alias{prepare_regressor_data}
\title{Prepare regressor data from vst object and cell attributes}
\usage{
prepare_regressor_data(vst_out, cell_attr)
}
\arguments{
\item{vst_out}{vst object containing model_str and model_pars_nonreg}

\item{cell_attr}{Data frame of cell attributes}
}
\value{
Matrix of regressor data
}
\description{
Helper function to create regressor data matrix, handling both regular
and non-regularized parameters if present
}
