% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_scale_binned}
\alias{robust_scale_binned}
\title{Robust scale using median and mad per bin}
\usage{
robust_scale_binned(y, x, breaks)
}
\arguments{
\item{y}{Numeric vector}

\item{x}{Numeric vector}

\item{breaks}{Numeric vector of breaks}
}
\value{
Numeric vector of scaled values
}
\description{
Put the values in \code{y} in bins based on the values in \code{x} and the
breaks defined in \code{breaks}. Apply robust scaling to \code{y} per bin.
}
