% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_gmean}
\alias{row_gmean}
\title{Geometric mean per row}
\usage{
row_gmean(x, eps = 1)
}
\arguments{
\item{x}{matrix of class \code{matrix} or \code{dgCMatrix}}

\item{eps}{small value to add to x to avoid log(0); default is 1}
}
\value{
geometric means
}
\description{
Geometric mean per row
}
