% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_gmean_grouped}
\alias{row_gmean_grouped}
\title{Geometric mean per row grouped by a factor}
\usage{
row_gmean_grouped(x, group, eps = 1)
}
\arguments{
\item{x}{matrix of class \code{dgCMatrix}}

\item{group}{factor to group the columns by (will be converted using \code{as.factor} and \code{droplevels})}

\item{eps}{small value to add to x to avoid log(0); default is 1}
}
\value{
matrix of geometric means
}
\description{
Geometric mean per row grouped by a factor
}
