% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_gmeans}
\alias{row_gmeans}
\title{Geometric mean per row}
\usage{
row_gmeans(x, eps = 1)
}
\arguments{
\item{x}{matrix}

\item{eps}{small value to add to x to avoid log(0); default is 1}
}
\value{
geometric means
}
\description{
Geometric mean per row
}
