% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_mean_grouped}
\alias{row_mean_grouped}
\title{Arithmetic mean per row grouped by a factor}
\usage{
row_mean_grouped(x, group)
}
\arguments{
\item{x}{matrix of class \code{dgCMatrix}}

\item{group}{factor to group the columns by (will be converted using \code{as.factor} and \code{droplevels})}
}
\value{
matrix of arithmetic means
}
\description{
Arithmetic mean per row grouped by a factor
}
