% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_progress_bar}
\alias{setup_progress_bar}
\title{Setup progress bar for batch processing}
\usage{
setup_progress_bar(n_items, bin_size, verbosity)
}
\arguments{
\item{n_items}{Total number of items to process}

\item{bin_size}{Size of each batch/bin}

\item{verbosity}{Verbosity level (0, 1, or 2)}
}
\value{
List with bin_ind (vector), pb (progress bar object or NULL), and max_bin
}
\description{
Setup progress bar for batch processing
}
