% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{umify_data}
\alias{umify_data}
\title{Transformation functions for umify}
\format{
A list of length two. The first element is a data frame with group, quantile and 
log-counts values. The second element is a vector of breaks to be used with cut to group
observations.
}
\usage{
umify_data
}
\description{
The functions have been trained on various public data sets and relate quantile
values to log-counts. Here the expected values at various points are given.
}
\keyword{datasets}
