% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_progress_bar}
\alias{update_progress_bar}
\title{Update progress bar}
\usage{
update_progress_bar(pb, i, verbosity)
}
\arguments{
\item{pb}{Progress bar object (can be NULL)}

\item{i}{Current iteration number}

\item{verbosity}{An integer specifying the verbosity level: 0 (silent), 1 (messages only), or 2 (messages and progress bars)}
}
\description{
Update progress bar
}
