% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{close_progress_bar}
\alias{close_progress_bar}
\title{Close progress bar}
\usage{
close_progress_bar(pb, verbosity)
}
\arguments{
\item{pb}{Progress bar object (can be NULL)}

\item{verbosity}{Verbosity level}
}
\description{
Close progress bar
}
