% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_residual_var}
\alias{get_residual_var}
\title{Return variance of residuals of regularized models}
\usage{
get_residual_var(
  vst_out,
  umi,
  residual_type = "pearson",
  res_clip_range = c(-sqrt(ncol(umi)), sqrt(ncol(umi))),
  min_variance = vst_out$arguments$min_variance,
  cell_attr = vst_out$cell_attr,
  bin_size = 256,
  verbosity = vst_out$arguments$verbosity
)
}
\arguments{
\item{vst_out}{The output of a vst run}

\item{umi}{The UMI count matrix that will be used}

\item{residual_type}{What type of residuals to return; can be 'pearson' or 'deviance'; default is 'pearson'}

\item{res_clip_range}{Numeric of length two specifying the min and max values the residuals will be clipped to; default is c(-sqrt(ncol(umi)), sqrt(ncol(umi)))}

\item{min_variance}{Lower bound for the estimated variance for any gene in any cell when calculating pearson residual; default is vst_out$arguments$min_variance}

\item{cell_attr}{Data frame of cell meta data}

\item{bin_size}{Number of genes to put in each bin (to show progress)}

\item{verbosity}{An integer specifying whether to show only messages (1), messages and progress bars (2) or nothing (0) while the function is running; default is 2}
}
\value{
A vector of residual variances (after clipping)
}
\description{
This never creates the full residual matrix and can be used to determine highly variable genes.
}
\examples{
\donttest{
vst_out <- vst(pbmc, return_cell_attr = TRUE)
res_var <- get_residual_var(vst_out, pbmc)
}

}
