% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{Peripheral Blood Mononuclear Cells (PBMCs)}
\format{A sparse matrix (dgCMatrix, see Matrix package) of molecule counts.
There are 12572 rows (genes) and 402 columns (cells).}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc
}
\description{
UMI counts for a subset of cells freely available from 10X Genomics
}
\keyword{datasets}
