% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_scale_binned}
\alias{robust_scale_binned}
\title{Robust scale using median and mad per bin}
\usage{
robust_scale_binned(y, x, breaks)
}
\arguments{
\item{y}{Numeric vector}

\item{x}{Numeric vector}

\item{breaks}{Numeric vector of breaks}
}
\value{
Numeric vector of scaled score
}
\description{
Robust scale using median and mad per bin
}
